==> src/01_name_items.ipynb <==
# Identify products

import pandas as pd
from pathlib import Path
import csd as c

## <span style="color:red"> Change vendor name in quotes

df = c.pull_cost_data(vendor_name='MERFISH UNITED')

## Pull all of vendor's products

df.head()

df.shape

df['prod'].drop_duplicates().shape

df.columns

df.to_clipboard(index=False)

## Check if other vendors supply some of the same products

products_series = df['prod'].drop_duplicates()

products_list = products_series.to_list()

products_CSS = ', '.join(products_list)

products_CSS

equivalents_df = c.pull_cost_data(prod=products_CSS)

equivalents_df

# get rid of rows with irrelevant vendors with PASSAIC METAL, PAMPCO, or GURTMAN in their names
equivalents_df_cleaned = equivalents_df.loc[(equivalents_df['name'].str.contains('PASSAIC METAL') == False) & 
                                            (equivalents_df['name'].str.contains('PAMPCO') == False) &
                                            (equivalents_df['name'].str.contains('GURTMAN') == False) ]

## <span style="color:red"> Change Vendor Name in Quotes

equivalents_other_vendors_df = equivalents_df_cleaned.loc[equivalents_df_cleaned['name'].str.contains('MERFISH UNITED') == False]

## <span style="color:red"> Change Vendor Name in Quotes

if equivalents_df_cleaned.shape[0] == 0:
    final_df = df
else:
    for item in equivalents_other_vendors_df['prod'].unique():
        item_df = equivalents_df_cleaned.loc[equivalents_df_cleaned['prod'] == item]
        item_t12m_sales_df = item_df.groupby(by=['prod','name'],as_index=False).agg({'t12m_sales':'sum'})
        max_sales = item_t12m_sales_df['t12m_sales'].max()
        if item_t12m_sales_df.loc[item_t12m_sales_df['name'] == 'MERFISH UNITED', 't12m_sales'].to_list()[0] == max_sales:
            rows_to_add_df = equivalents_other_vendors_df.loc[equivalents_other_vendors_df['prod'] == item]
            df = pd.concat([df,rows_to_add_df])
        else:
            df = df
    final_df = df

## Save Product Info

## <span style="color:red"> Change the vendor name in last part of file path

loc = Path.cwd().parent / 'data' / 'merfish_product_info.csv'
final_df.to_csv(loc, index=False)
