==> src/01_pull_invoices.ipynb <==
# Pricing Project
Develop a pricing program by which I stratify products.

import pandas as pd
from pathlib import Path
import csd as c
from functions4 import pull_raw_data

pd.set_option('display.max_columns', 100)

### Pull Invoice Data

start_date = '2022-02-01'
end_date = '2023-01-31'

df = pull_raw_data(start_date, end_date)

df.head()

### Export to Excel

loc = Path.cwd().parent / 'data' / 'generated' / 'invoice_data_0_raw.csv'
df.to_csv(loc, index=False)

df.groupby(['month']).sum().sort_values('month', ascending=False)[['netamt', 'GP$', 'xcost_adj', 'units']]

df.shape
