==> src/02_identify_exclusions.ipynb <==
# Pricing Project
## Identify Exclusions
Need to identify exclusions for:
* Outliers
* Customers
* Items

import pandas as pd
from pathlib import Path

pd.set_option('display.max_columns', 100)

loc = Path.cwd().parent / 'data' / 'generated' / 'invoice_data_0_raw.csv'
inv_df = pd.read_csv(loc)

inv_df.head()

# add Exclusion column to raw file
inv_df['Exclusion'] = None

## Items
* Nonstock
* Special Order
* Labor Codes

# Remove any Nonstock Items
null_criteria = inv_df['Exclusion'].isnull()
ns_criteria = inv_df['prod_type'] == 'nonstock'
inv_df.loc[null_criteria & ns_criteria, 'Exclusion'] = 'Item - nonstock'

# Remove any Special Order Items
null_criteria = inv_df['Exclusion'].isnull()
so_criteria = inv_df['prod_type'] == 'special order'
inv_df.loc[null_criteria & so_criteria, 'Exclusion'] = 'Item - special order'

# Remove Labor Code Items
null_criteria = inv_df['Exclusion'].isnull()
labor_criteria = inv_df['prodcat'] == 8888
inv_df.loc[null_criteria & labor_criteria, 'Exclusion'] = 'Item - labor'

inv_df['prod_type'].unique()

## Customers
* Accounts that don't have custname filled in, assuming these are inactive
* Customer with XXX in the name
* Any customers that appear in the ../data/inputs/customer_exclusion.xlsx file

# fill in missing customer names with 'Inactive'
inv_df['custname'] = inv_df['custname'].fillna('INACTIVE')

# Remove any inactive customers
null_criteria = inv_df['Exclusion'].isnull()
inactive_acct_criteria = inv_df['custname'].str.contains('INACTIVE')
inv_df.loc[null_criteria & inactive_acct_criteria, 'Exclusion'] = 'Customer - inactive acct'

# Remove any customers with 'XXX' in the name
null_criteria = inv_df['Exclusion'].isnull()
xxx_acct_criteria = inv_df['custname'].str.contains('XXX')
inv_df.loc[null_criteria & xxx_acct_criteria, 'Exclusion'] = 'Customer - xxx acct'

# Remove any customers in the ../data/customer_exclusions.xlsx file
loc = Path.cwd().parent / 'data' / 'inputs' / 'customer_exclusions.xlsx'
cust_ex_df = pd.read_excel(loc, engine='openpyxl')
cust_exclusion_map = {i[0]:i[1] for i in cust_ex_df.to_dict('split')['data']}
inv_df['Exclusion'] = inv_df['Exclusion'].combine_first(inv_df['custno'].map(cust_exclusion_map))

## Outliers
* Quantity Shipped < 0
* Unit Cost = 0
* Unit Price = 0
* Net Amount <= 0
* Gross Profit Dollars <= 0
* Gross Profit Percent < 0.5% or > 80%

# Remove any lines with negative quantity shipped
null_criteria = inv_df['Exclusion'].isnull()
qty_ship_criteria = inv_df['units'] < 0
inv_df.loc[null_criteria & qty_ship_criteria, 'Exclusion'] = 'Outlier - negative quantity'

# Remove any lines with zero unit cost
null_criteria = inv_df['Exclusion'].isnull()
unit_cost_criteria = inv_df['unitcost'] == 0
inv_df.loc[null_criteria & unit_cost_criteria, 'Exclusion'] = 'Outlier - zero cost'

# Remove any lines with zero unit price
null_criteria = inv_df['Exclusion'].isnull()
unit_price_criteria = inv_df['unitprice'] == 0
inv_df.loc[null_criteria & unit_price_criteria, 'Exclusion'] = 'Outlier - zero price'

# Remove any lines with net amount less than or equal to zero
null_criteria = inv_df['Exclusion'].isnull()
unit_price_criteria = inv_df['netamt'] <= 0
inv_df.loc[null_criteria & unit_price_criteria, 'Exclusion'] = 'Outlier - zero net amount'

# Remove any lines with negative gross profit dollars
null_criteria = inv_df['Exclusion'].isnull()
gp_dollar_criteria = inv_df['GP$'] <= 0
inv_df.loc[null_criteria & gp_dollar_criteria, 'Exclusion'] = 'Outlier - negative margin dollars'

# Remove any lines with negative gross profit dollars
null_criteria = inv_df['Exclusion'].isnull()
gp_pct_low_criteria = inv_df['Margin'] < .005
gp_pct_high_criteria = inv_df['Margin'] > .8
inv_df.loc[null_criteria & (gp_pct_low_criteria | gp_pct_high_criteria), 'Exclusion'] = 'Outlier - high/low margin pct'

## Max Gurtman: whse = 601 or 602

# Remove any lines with whse = 601
null_criteria = inv_df['Exclusion'].isnull()
Max_Gurtman_601_criteria = inv_df['whse'] == 601
inv_df.loc[null_criteria & Max_Gurtman_601_criteria, 'Exclusion'] = 'Warehouse - 601'

# Remove any lines with whse = 602
null_criteria = inv_df['Exclusion'].isnull()
Max_Gurtman_602_criteria = inv_df['whse'] == 602
inv_df.loc[null_criteria & Max_Gurtman_602_criteria, 'Exclusion'] = 'Warehouse - 602'

## Summary

inv_df.loc[inv_df['prod_type'] == 'nonstock'].groupby(by='vendno', as_index=False).agg({'netamt':'sum'}).to_clipboard(index=False)

inv_df['Exclusion'].value_counts()

inv_df[['Exclusion','netamt']].groupby(by='Exclusion').sum('netamt')

inv_df[['Exclusion']].isna().any(axis=1).sum()

## Save to csv

loc = Path.cwd().parent / 'data' / 'generated' / 'invoice_data_1_exclusions_labeled.csv'
inv_df.to_csv(loc, index=False)

inv_df.loc[inv_df['Exclusion'] == 'Customer - inactive acct']['netamt'].sum()

inv_df.loc[inv_df['Exclusion'] == 'Warehouse - 602']['netamt'].sum()

sum(inv_df['netamt'])
