==> src/02_name_PT_7_PRs.ipynb <==
import pandas as pd
from pathlib import Path
import csd as c

## Find counter prices

df = c.pull_counter_price()

df.head()

## <span style="color:red"> Change vendor name in last part of file path

loc = Path.cwd().parent / 'data' / 'merfish_product_info.csv'
products_df = pd.read_csv(loc)

products_df.head()

PT_7_PRs_df = pd.DataFrame()

for item in products_df['prod']:
    partial_df = df.loc[df['prod'] == item]
    PT_7_PRs_df = pd.concat([PT_7_PRs_df,partial_df])

PT_7_PRs_df.head()

PT_7_PRs_df.shape

## <span style="color:red"> Change vendor name in last part of vendor path

loc = Path.cwd().parent / 'data' / 'merfish_PT_7_PRs.csv'
PT_7_PRs_df.to_csv(loc, index=False)
