==> src/03_create_clean_invoice_file.ipynb <==
# Curate in-scope invoice data set
Keep only data that is not labeled as excluded for any reason, and add any necessary labels.

Labels:
* Warehouse demographics (division, warehouse name)

import pandas as pd
from pathlib import Path

loc = Path.cwd().parent / 'data' / 'generated' / 'invoice_data_1_exclusions_labeled.csv'
inv_df_raw = pd.read_csv(loc)

# keep only clean data
inv_df_clean = inv_df_raw.loc[inv_df_raw['Exclusion'].isnull()].copy()

inv_df_clean.to_clipboard(index=False)

# merge in warehouse demographics
loc = Path.cwd().parent / 'data' / 'inputs' / 'warehouse_demographics.xlsx'
whse_demo_df = pd.read_excel(loc, engine='openpyxl')

inv_df_clean = inv_df_clean.merge(whse_demo_df, how='left', on='whse', validate='m:1')

inv_df_clean.head(3)

sum(inv_df_clean['netamt'])

loc = Path.cwd().parent / 'data' / 'generated' / 'invoice_data_2_in-scope.csv'
inv_df_clean.to_csv(loc, index=False)

inv_df_clean.groupby(by=['item'],as_index=False).size()
