==> src/03_name_sales.ipynb <==
# Identify sales

import pandas as pd
from pathlib import Path
import csd as c

## <span style="color:red"> Update Month in parentheses below as you update the auxiliary folder monthly

loc = Path.cwd().parent.parent / 'auxiliary (january)' / 'data' / 'generated' / 'invoice_data_2_in-scope.csv'
inv_df = pd.read_csv(loc)

## All invoice lines of sales

## <span style="color:red"> Change vendor number(s)

loc = Path.cwd().parent / 'data' / 'merfish_product_info.csv'
prod_info_df = pd.read_csv(loc)

unique_prod_df = prod_info_df[['prod']].drop_duplicates()

sales_df = unique_prod_df.merge(inv_df, how='inner', left_on='prod', right_on='item').drop(columns=['prod'])

sales_df.to_clipboard(index=False)

sales_df.head()

sales_df.columns

## Total sales by item and total customers

"""
group sales in two stages:
    first stage by 'item' and 'custno' with 'units' aggregated by sum
    second stage by 'item' aggregating 'custno' by count and 'units' by sum
the two stages of process let us count # of customers and get total sales of each item
"""
agg_sales_pre_df = sales_df.groupby(by=['item','itemdesc','custno'],as_index=False).agg({'units':'sum','netamt':'sum','xcost_adj':'sum'})

agg_sales_pre_df.to_clipboard(index=False)

agg_sales_pre_df

agg_sales_df = agg_sales_pre_df.groupby(by=['item','itemdesc'],as_index=False).agg({'custno':'count','units':'sum','netamt':'sum','xcost_adj':'sum'}).rename(columns={'custno':'total cust\'s','units':'total units'})

agg_sales_df['total margin'] = round( 100*(agg_sales_df['netamt'] - agg_sales_df['xcost_adj']) / agg_sales_df['netamt'], 2)

agg_sales_df = agg_sales_df.drop(columns=['netamt','xcost_adj'])

agg_sales_df.to_clipboard(index=False)

## Sales on overrides

overrides_df = sales_df.loc[sales_df['priceorigcd'] == 'O']

overrides_df.drop(columns=['ordersuf','lineno','unit','unitconv','month','Exclusion','cat_descrip','whse_name','division'])

agg_overrides_df = overrides_df.groupby(by=['item'],as_index=False).agg({'units':'sum'}).rename(columns={'units':'units overriden'})

agg_overrides_df

## Total PT 7 sales (no overrides)

## <span style="color:red"> Change vendor name in last part of file path

loc = Path.cwd().parent / 'data' / 'merfish_PT_7_PRs.csv'
PT_7_PRs = pd.read_csv(loc)

PT_7_PRs = PT_7_PRs.drop_duplicates()

PT_7_sales_by_PR_df = sales_df.merge(PT_7_PRs[['prod','pdrecno']], how='inner', left_on=['item','pdrecno'], right_on=['prod','pdrecno'])

PT_7_sales_by_PR_df.head()

PT_7_sales_by_PR_df.shape

PT_7_sales_by_PR_df['margin by replcost'] = (PT_7_sales_by_PR_df['netamt'] - PT_7_sales_by_PR_df['replcost']*PT_7_sales_by_PR_df['units']) / PT_7_sales_by_PR_df['netamt']

# drop unnecessary columns
PT_7_sales_by_PR_df = PT_7_sales_by_PR_df.drop(columns=['orderno','ordersuf','lineno','unit','unitconv','returnfl','prod_type','prodcat','cat_descrip','rowpointer','Exclusion'])

# inspect output
PT_7_sales_by_PR_df.to_clipboard(index=False)

PT_7_no_overs_df = PT_7_sales_by_PR_df.loc[ PT_7_sales_by_PR_df['priceorigcd'] == '7']

# inspect output
PT_7_no_overs_df.to_clipboard(index=False)

"""
group sales in two stages:
    first stage by 'item' and 'custno' with 'units' aggregated by sum
    second stage by 'item' aggregating 'custno' by count and 'units' by sum
the two stages of process let us count # of customers and get total sales of each item
"""
PT_7_no_overs_agg_pre_df = PT_7_no_overs_df.groupby(by=['item','custno'],as_index=False).agg({'units':'sum'})

PT_7_no_overs_agg_pre_df.to_clipboard(index=False)

PT_7_no_overs_agg_pre_df.loc[PT_7_no_overs_agg_pre_df.duplicated(subset=['item','custno'])]

PT_7_no_overs_agg_df = PT_7_no_overs_agg_pre_df.groupby(by=['item'],as_index=False).agg({'custno':'count','units':'sum'}).rename(columns={'custno':'PT 7 cust\'s', 'units':'PT 7 units'})

PT_7_no_overs_agg_df

# make cumulative percentage stats
PT_7_no_overs_quartiles_df = PT_7_no_overs_agg_pre_df[['item','units']].groupby('item').quantile(q=[.5,1],interpolation='midpoint')

PT_7_no_overs_quartiles_df.head()

PT_7_no_overs_quartiles_df = PT_7_no_overs_quartiles_df.reset_index()

# data frame of quartiles of sales per customer
quartile_df = pd.DataFrame([ [prod,tuple(PT_7_no_overs_quartiles_df.loc[PT_7_no_overs_quartiles_df['item'] == prod]['units'].tolist())] for prod in PT_7_no_overs_quartiles_df['item'].unique()], columns=['item','PT 7 quartiles'])

quartile_df

# Standard cost, multiplier, and standard price

## <span style="color:red"> Change the vendor name at end of file path

loc = Path.cwd().parent / 'data' / 'merfish_product_info.csv'
cost_data_df = pd.read_csv(loc)

cost_data_df.columns

## <span style="color:red"> Change vendor number(s)

# pick out cost data for our vendor
specific_cost_data_df = cost_data_df.loc[ ((cost_data_df['vendno'] == 18821) | (cost_data_df['vendno'] == 86125)) ]

specific_cost_data_df_cleaned = specific_cost_data_df[['prod','descrip_1','descrip_2','stndcostdt','stndcost','prcmult_1','stndprice','vendno']].sort_values(by=['prod','stndcostdt','stndcost']).drop_duplicates(subset=['prod'],keep='last')

specific_cost_data_df_cleaned['stndmargin'] = round((specific_cost_data_df_cleaned['prcmult_1']*100 - 100) / specific_cost_data_df_cleaned['prcmult_1'], 2)

specific_cost_data_df_cleaned

## Put everything together

merge_1_df = agg_sales_df.merge(agg_overrides_df, how='left', on='item')

merge_2_df = merge_1_df.merge(PT_7_no_overs_agg_df, how='left', on='item')

merge_3_df = merge_2_df.merge(quartile_df, how='left', on='item')

merge_4_df = merge_3_df.merge(specific_cost_data_df_cleaned, how='outer', left_on='item', right_on='prod')

final_df = merge_4_df.drop(columns='item')

final_df.columns

final_df.head()

final_df['new stndcost'] = None
final_df['stndprice no adjust.'] = None
final_df['new stndprice'] = None
final_df['new stndmargin'] = None
final_df['move rate'] = None
final_df['margin change'] = None

final_df = final_df[['prod','descrip_1', 'descrip_2', 'total cust\'s', 'total units', 'total margin', 'units overriden',
                     'PT 7 cust\'s', 'PT 7 units', 'PT 7 quartiles', 'stndcost', 'stndprice', 'stndmargin', 'new stndcost',
                     'stndprice no adjust.', 'move rate', 'new stndmargin', 'new stndprice', 'margin change']]

## Save records of sales

## <span style="color:red"> Change vendor name at end of file path

loc = Path.cwd().parent / 'data' / 'merfish_sales_record.csv'
final_df.to_csv(loc, index=False)
